+++ design-system/stories/Button/HorizontalCardButton.stories.tsx
@@ -1,5 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import HorizontalCardButton from '../../ui/Button/HorizontalCardButton';
+
 type Story = StoryObj<typeof meta>;
 
 const meta = {
